fixing specifications, endpoints still 404

+++ src/main/java/Services/BookingService.java
@@ -5,34 +5,33 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.swing.text.html.Option;
+import java.awt.print.Book;
 import java.time.Duration;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class BookingService {
-    /*
-    Most likely need property repo/service
-     */
+    // Most likely need property repo/service
+    @Autowired
     BookingRepo bookingRepo;
 
+    /*
     @Autowired
     public BookingService(BookingRepo bookingRepo){
         this.bookingRepo = bookingRepo;
     }
+    */
 
     //needs to throw exceptions if outside of date range
     //throw exception if property is already booked
     public Booking addBooking(Booking booking) {
-        if (booking.getEnd_date().isBefore(booking.getStart_date())) {
-            return null;
-        }
-
-        int attempt = booking.getBooking_id();
-        Optional<Booking> bookingAttempt = bookingRepo.findById(attempt);
-        if (bookingAttempt.isPresent()) {
+        if((booking.getStatus().equals("confirmed")) || (booking.getEnd_date().isBefore(booking.getStart_date()))){
             return null;
         } else {
+            Long attempt = booking.getBooking_id();
+            Optional<Booking> bookingAttempt = bookingRepo.findById(attempt);
+            booking.setStatus("confirmed");
             return bookingRepo.save(booking);
         }
     }
@@ -51,16 +50,19 @@ public class BookingService {
         @param: daterange
         @param: Size
         @param: Owner
-     */
+
+    //Need info from property table
     public Booking getBookingByGuest(int guest_id){
         //need different findby
         Optional<Booking> optionalBooking = bookingRepo.findById(guest_id);
         Booking booking = optionalBooking.get();
         return booking;
     }
+     */
 
-    public Booking getBookingByProperty(int property_id){
-        Optional<Booking> optionalBooking = bookingRepo.findById(property_id);
+    //I don't think I actually need this lol
+    public Booking getBookingByProperty(Long property_id){
+        Optional<Booking> optionalBooking = bookingRepo.findByProperty_id(property_id);
         Booking booking = optionalBooking.get();
         return booking;
     }
@@ -75,8 +77,12 @@ public class BookingService {
     }
      */
 
-    public void deleteBooking(Integer booking_id){
-        bookingRepo.deleteById(booking_id);
+    public void deleteBooking(Long booking_id) {
+        Optional<Booking> optionalBooking = bookingRepo.findByProperty_id(booking_id);
+        if (optionalBooking.isPresent()) {
+            optionalBooking.get().setStatus("cancelled");
+            bookingRepo.deleteById(booking_id);
+        }
     }
 
 }
